+++ public/vocables/modules/shared/subGraph.js
@@ -9,37 +9,134 @@ var SubGraph = (function () {
         if (!options) {
             options = {};
         }
-        var classesMap = {};
-        var allClasses = [];
-        var allRestrictionProperties = [];
+        var nClasses = 1
+        var allRestrictions = {}
+        var allClasses = {}
+        var uniqueRestrictions = {}
 
         var fromStr = Sparql_common.getFromStr(sourceLabel);
+
+        var currentClasses = [baseClassId]
+
         async.series(
             [
+
                 function (callbackSeries) {
-                    OntologyModels.registerSourcesModel(sourceLabel, null, function (err, result) {
-                        callbackSeries(err);
-                    });
+
+
+                    async.whilst(
+                        function (callbackTest) {
+                            return nClasses > 0;
+                        },
+
+                        function (callbackWhilst) {
+                            var filter = Sparql_common.setFilter("s", currentClasses, null, {values: true});
+
+
+                            var query = "PREFIX dexp: <http://totalenergies/resources/tsf/ontology/dexpi-process/specific/>\n" +
+                                "PREFIX owl: <http://www.w3.org/2002/07/owl#>PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>SELECT distinct *  FROM   <http://totalenergies/resources/tsf/ontology/dexpi-process/specific/>    FROM   <http://totalenergies/resources/tsf/ontology/dexpi-process/generic/>  WHERE {\n" +
+                                "  ?s rdfs:subClassOf+ ?o. ?o rdf:type ?type " +
+                                filter +
+                                "optional { ?o owl:onProperty ?property. ?o owl:someValuesFrom|owl:onClass ?targetClass  optional { ?o ?cardinalityType  ?cardinalityValue. filter (?cardinality in (owl:minCardinality,owl:maxCardinality,owl:cardinality))}}\n" +
+                                "  } limit 10000"
+                            self.query(sourceLabel, query, function (err, result) {
+                                if (err) {
+                                    return alert(err);
+                                }
+                                currentClasses = [];
+                                result.results.bindings.forEach(function (item) {
+                                    if (!item.type) {
+                                        return;
+                                    }
+
+                                    if (item.type.value.endsWith("Class")) {
+                                        if (!allClasses[item.o.value]) {
+                                            currentClasses.push(item.o.value)
+                                            allClasses[item.o.value] = {}
+                                        }
+                                        if (!allClasses[item.s.value]) {
+                                            currentClasses.push(item.s.value)
+
+                                        }
+                                        allClasses[item.s.value] = {superClass: item.o.value}
+                                    } else if (item.type.value.endsWith("Restriction")) {
+
+                                        if (!allRestrictions[item.s.value]) {
+                                            allRestrictions[item.s.value] = {}
+                                        }
+
+                                        var obj = {
+                                            property: item.property ? item.property.value : null,
+                                            targetClass: item.targetClass ? item.targetClass.value : null,
+                                            cardinalityType: item.cardinalityType ? item.cardinalityType.value : null,
+                                            cardinalityValue: item.cardinalityValue ? item.cardinalityValue.value : null,
+
+
+                                        }
+                                        if (obj.targetClass.endsWith("Volume")) {
+                                            var x = 1
+                                        }
+
+                                        if (item.targetClass) {
+                                            if (!allClasses[item.targetClass.value]) {
+                                                allClasses[item.targetClass.value] = {}
+                                                currentClasses.push(item.targetClass.value)
+                                            }
+                                        }
+
+                                        if (!allRestrictions[item.s.value][obj.property]) {
+                                            allRestrictions[item.s.value][obj.property] = []
+                                        }
+                                        if (!uniqueRestrictions[obj.property + obj.targetClass + obj.cardinalityType]) {
+                                            uniqueRestrictions[obj.property + obj.targetClass + obj.cardinalityType] = 1
+                                            allRestrictions[item.s.value][obj.property].push(obj);
+                                        }
+
+
+                                    }
+                                });
+
+
+                                nClasses = currentClasses.length;
+                                callbackWhilst();
+                            });
+
+                        },
+                        function (err) {
+
+                            callbackSeries();
+                        })
+
+
                 },
 
                 function (callbackSeries) {
-                    options.restrictions = true;
-                    OntologyModels.getAllowedPropertiesBetweenNodes(sourceLabel, [baseClassId], null, options, function (err, result) {
-                        //  OntologyModels.getObjectPropertiesFromRestrictions(sourceLabel, [baseClassId], null, options, function (err, result) {
+
+                    var uris = Object.keys(allClasses).concat(Object.keys(allRestrictions))
+                    var filter = Sparql_common.setFilter("s", uris)
+
+                    var query = "PREFIX dexp: <http://totalenergies/resources/tsf/ontology/dexpi-process/specific/>\n" +
+                        "PREFIX owl: <http://www.w3.org/2002/07/owl#>PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>SELECT distinct *  FROM   <http://totalenergies/resources/tsf/ontology/dexpi-process/specific/>    FROM   <http://totalenergies/resources/tsf/ontology/dexpi-process/generic/>  WHERE {\n" +
+                        "  ?s rdfs:label ?sLabel. " +
+                        filter + "} limit 10000"
+                    self.query(sourceLabel, query, function (err, result) {
                         if (err) {
-                            return callbackSeries(null, {});
+                            return alert(err);
                         }
-                        allRestrictionProperties = result.constraints.noConstraints;
+                        result.results.bindings.forEach(function (item) {
+                            if (allClasses[item.s.value]) {
+                                allClasses[item.s.value].label = item.sLabel.value
+                            }
+
+                        })
                         return callbackSeries();
-                    });
-                },
+                    })
+                }
+
+
             ],
             function (err) {
-                return callback(err, {
-                    classes: allClasses,
-                    restrictions: allRestrictionProperties,
-                    classesMap: classesMap,
-                });
+                return callback(err, {classes: allClasses, restrictions: allRestrictions});
             }
         );
     };
@@ -129,67 +226,83 @@ var SubGraph = (function () {
         });
     };
 
-    self.getSubGraphShaclTriples = function (sourceLabel, classUri, options, callback) {
+    self.getSubGraphShaclTriples = function (sourceLabel, _classUri, options, callback) {
         if (!options) {
             options = {};
         }
-        if (!classUri) {
-            classUri = "http://tsf/resources/ontology/DEXPIProcess_gfi_2/TransportingFluidsActivity";
+
+
+        self.classUri = _classUri
+        if (!self.classUri) {
+            self.classUri = "http://tsf/resources/ontology/DEXPIProcess_gfi_2/TransportingFluidsActivity";
         }
 
         Shacl.initSourceLabelPrefixes(sourceLabel);
 
-        self.getSubGraphResources(sourceLabel, classUri, options, function (err, result) {
-            var classes = {};
-            for (var property in result.restrictions) {
-                var items = result.restrictions[property];
-                if (!Array.isArray(items)) items = [items];
-                items.forEach(function (item) {
-                    if (!classes[item.domain]) {
-                        classes[item.domain] = [];
-                    }
-                    item.property = property;
-                    classes[item.domain].push(item);
-                });
+        self.getSubGraphResources(sourceLabel, self.classUri, options, function (err, result) {
+            //  var resources = result.classes.concat(result.restrictions);
+            var classesMap = result.classes;
+            var restrictionsMap = result.restrictions
+            // return;
+
+
+
+
+            var usedClasses = {}
+            usedClasses[self.classUri] = 1
+            for (var classUri2 in restrictionsMap) {
+                for (var property in restrictionsMap[classUri2]) {
+                    restrictionsMap[classUri2][property].forEach(function (item) {
+                        if (item.targetClass) {
+                            usedClasses[item.targetClass] = 1
+                        }
+
+                    })
+                }
             }
 
             var allSahcls = "";
-            for (var classUri in classes) {
-                var classRestrictions = classes[classUri];
+            for (var classUri2 in usedClasses) {
+                var item = classesMap[classUri2];
+
+                var classRestrictions = restrictionsMap[classUri2]
                 var shaclProperties = [];
+                if (classRestrictions) {
 
-                classRestrictions.forEach(function (restriction) {
-                    if (!restriction.property || !restriction.range) {
-                        return;
+                    for (var property in classRestrictions) {
+
+                        classRestrictions[property].forEach(function (restriction) {
+                            if (!restriction.property || !restriction.targetClass) {
+                                return;
+                            }
+                            var count = -1;
+                            if (restriction.cardinalityValue) {
+                                count = restriction.cardinalityValue
+                            }
+
+
+                            var propStr = Shacl.uriToPrefixedUri(restriction.property);
+                            var rangeStr = Shacl.uriToPrefixedUri(restriction.targetClass);
+                            var property = " sh:path " + propStr + " ;\n";
+                            if (count > -1) {
+                                property += "        sh:minCount " + count + " ;";
+                            }
+                            //  "        sh:maxCount " + count + " ;" +
+                            property += "        sh:node " + rangeStr + " ;";
+
+                            shaclProperties.push(property);
+                        });
+
                     }
-                    var count = -1;
-                    if (restriction.cardinality) {
-                        count = Sparql_common.getIntFromTypeLiteral(restriction.cardinality);
-                    }
-
-                    if (restriction.property.endsWith("Quality")) {
-                        return;
-                    }
-                    var propStr = Shacl.uriToPrefixedUri(restriction.property);
-                    var rangeStr = Shacl.uriToPrefixedUri(restriction.range);
-                    var property = " sh:path " + propStr + " ;\n";
-                    if (count > -1) {
-                        property += "        sh:minCount " + count + " ;";
-                    }
-                    //  "        sh:maxCount " + count + " ;" +
-                    property += "        sh:node " + rangeStr + " ;";
-
-                    shaclProperties.push(property);
-                });
-
-                var domain = Shacl.uriToPrefixedUri(classUri);
-                var shaclStr = Shacl.getShacl(classUri, null, shaclProperties);
-
-                allSahcls += "\n" + shaclStr;
+                    var domain = Shacl.uriToPrefixedUri(classUri2);
+                    var shaclStr = Shacl.getShacl(domain, null, shaclProperties);
+                    allSahcls += shaclStr;
+                }
             }
 
-            allSahcls = Shacl.getPrefixes() + allSahcls;
 
+            var prefixes = Shacl.getPrefixes();
+            allSahcls = prefixes + "\n" + allSahcls+".";
             var payload = {
                 turtle: allSahcls,
             };
@@ -205,14 +318,16 @@ var SubGraph = (function () {
                     if (data.result && data.result.indexOf("Error") > -1) {
                         return callback(data.result);
                     }
-                    return callback(null, data.triples);
+                    return callback(null, {triples: data.triples, shacl:allSahcls, classesMap: classesMap});
                     //  callback(null, data);
                 },
                 error(err) {
                     callback(err.responseText);
                 },
             });
-        });
+        })
+
+
     };
 
     self.getResourcesPredicates = function (sourceLabel, resources, action, options, callback) {
@@ -281,7 +396,8 @@ var SubGraph = (function () {
             classUri = "http://tsf/resources/ontology/DEXPIProcess_gfi_2/TransportingFluidsActivity";
         }
 
-        self.instantiateSubGraph(sourceLabel, classUri, function (err, result) {});
+        self.instantiateSubGraph(sourceLabel, classUri, function (err, result) {
+        });
 
         return;
         self.getSubGraphResources(sourceLabel, classUri, options, function (err, result) {
